+++ cmd/root.go
@@ -12,7 +12,7 @@ import (
 	"github.com/wildcard/gh-code-review/internal/output"
 )
 
-var version = "0.1.0"
+var version = "0.1.1"
 
 type app struct {
 	exitCode int